Add `has_bokeh` function, for pages which glue but don’t produce figures
For pages which only glue Bokeh figures produced on other pages, this 
adds the "has_bokeh" metadata to a cell, thus faking the presence of a 
Bokeh figure and triggering MyST-NB-Bokeh to add the necessary JS to 
that page.

Usage: `from myst_nb_bokeh import has_bokeh; has_bokeh()`

There are probably cleaner ways to do this. Also, ideally we would just 
detect that a Bokeh figure has been glue, and not require adding code to 
the notebook. (Especially relevant for notebooks that otherwise wouldn’t 
need to be executed at all.)